+++ apps/fluent-tester/src/E2E/Persona/specs/Persona.spec.macos.ts
@@ -1,23 +1,21 @@
-import NavigateAppPage from '../../common/NavigateAppPage.macos';
-import PersonaPageObject from '../pages/PersonaPageObject.macos';
+import NavigateAppPage from '../../common/NavigateAppPage';
+import PersonaPageObject from '../pages/PersonaPageObject';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('Persona Testing Initialization', function () {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy();
   });
 
-  it('Click and navigate to Persona test page', () => {
-    /* Scroll to component test page button in scrollview if not already visible*/
-    PersonaPageObject.scrollToComponentButton();
-    PersonaPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
+  it('Click and navigate to Persona test page', async () => {
+    await PersonaPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
-    NavigateAppPage.clickAndGoToPersonaPage();
-    PersonaPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await NavigateAppPage.clickAndGoToPersonaPage();
+    await PersonaPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(PersonaPageObject.isPageLoaded()).toBeTruthy();
+    await expect(await PersonaPageObject.isPageLoaded()).toBeTruthy();
   });
 });