+++ apps/fluent-tester/src/E2E/common/NativeTesting/specs/NativeTesting.spec.win.ts
@@ -1,11 +1,11 @@
-import NavigateAppPage from '../../../common/NavigateAppPage.win';
+import NavigateAppPage from '../../NavigateAppPage';
 import NativeTestingPageObject from '../pages/NativeTestingPageObject.win';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../../common/consts';
 
 describe('Native Safety Check Testing Initialization', () => {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
   });
 
   // The ScrollView in testing is the one that contains all the buttons that navigate to each component's
@@ -14,15 +14,15 @@ describe('Native Safety Check Testing Initialization', () => {
   // We need to make sure this ScrollView:
   //    1) Continues to parent all the test buttons
   //    2) Keeps testID='SCROLLVIEW_TEST_ID'
-  it('Validate the ScrollView containing all the navigational buttons exists', () => {
-    NativeTestingPageObject.waitForScrollViewDisplayed(PAGE_TIMEOUT);
-    expect(NativeTestingPageObject.doesScrollViewParentExist()).toBeTruthy();
+  it('Validate the ScrollView containing all the navigational buttons exists', async () => {
+    await NativeTestingPageObject.waitForScrollViewDisplayed(PAGE_TIMEOUT);
+    await expect(await NativeTestingPageObject.doesScrollViewParentExist()).toBeTruthy();
   });
 
   // In addition to existing, we want to ensure the Children of the ScrollView stay intact. The children are the buttons that
   // navigate to each test page. Validate these children exist with the proper testId format.
-  it('Validate the children of ScrollView are the buttons that navigate to each test page with proper automationId', () => {
-    NativeTestingPageObject.waitForScrollViewDisplayed(PAGE_TIMEOUT);
-    expect(NativeTestingPageObject.validateScrollViewChildren()).toBeTruthy();
+  it('Validate the children of ScrollView are the buttons that navigate to each test page with proper automationId', async () => {
+    await NativeTestingPageObject.waitForScrollViewDisplayed(PAGE_TIMEOUT);
+    await expect(await NativeTestingPageObject.validateScrollViewChildren()).toBeTruthy();
   });
 });