+++ apps/fluent-tester/src/TestComponents/Common/MenuPicker.windows.tsx
@@ -1,14 +1,14 @@
 export * from './MenuPicker.desktop';
 import * as React from 'react';
 import { Button, Text, View, StyleSheet } from 'react-native';
-import { MenuPickerProps, collectionItem } from './MenuPicker';
+import { MenuPickerProps, CollectionItem } from './MenuPicker';
 
 export const MenuPicker: React.FunctionComponent<MenuPickerProps> = (props: MenuPickerProps) => {
   const { prompt, onChange, collection } = props;
   return (
     <View style={menuPickerStyles.container}>
       <Text style={menuPickerStyles.prompt}>{prompt}</Text>
-      {collection.map((item: collectionItem, index: number) => (
+      {collection.map((item: CollectionItem, index: number) => (
         <Button onPress={() => onChange(item.value, index)} key={index} title={item.label} />
       ))}
     </View>