+++ packages/components/RadioGroup/CHANGELOG.md
@@ -1,8 +1,47 @@
 # Change Log - @fluentui-react-native/radio-group
 
-This log was last generated on Fri, 26 Aug 2022 18:15:17 GMT and should not be manually modified.
+This log was last generated on Fri, 09 Sep 2022 21:53:29 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.15.12
+
+Fri, 09 Sep 2022 21:53:29 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.10.10
+
+## 0.15.11
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.10.9
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+- Bump @fluentui-react-native/pressable to v0.9.9
+- Bump @fluentui-react-native/text to v0.14.2
+
+## 0.15.10
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.12.2
+- Bump @fluentui-react-native/focus-zone to v0.10.8
+- Bump @fluentui-react-native/interactive-hooks to v0.18.2
+- Bump @fluentui-react-native/pressable to v0.9.8
+- Bump @fluentui-react-native/text to v0.14.1
+
+## 0.15.9
+
+Mon, 29 Aug 2022 22:11:28 GMT
+
+### Patches
+
+- Fix RTL for RadioButton and Checkbox (67026167+chiuam@users.noreply.github.com)
 
 ## 0.15.8