+++ packages/experimental/Switch/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log - @fluentui-react-native/switch
 
-This log was last generated on Fri, 26 Aug 2022 18:15:17 GMT and should not be manually modified.
+This log was last generated on Tue, 06 Sep 2022 18:22:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.4
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.14.2
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+
+## 0.6.3
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.14.1
+- Bump @fluentui-react-native/framework to v0.8.1
+- Bump @fluentui-react-native/interactive-hooks to v0.18.2
 
 ## 0.6.2