+++ packages/experimental/Tabs/CHANGELOG.md
@@ -1,8 +1,39 @@
 # Change Log - @fluentui-react-native/experimental-tabs
 
-This log was last generated on Fri, 05 Aug 2022 23:22:38 GMT and should not be manually modified.
+This log was last generated on Fri, 09 Sep 2022 21:53:29 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.12
+
+Fri, 09 Sep 2022 21:53:29 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.10.10
+
+## 0.5.11
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+- Bump @fluentui-react-native/focus-zone to v0.10.9
+- Bump @fluentui-react-native/text to v0.14.2
+- Bump @fluentui-react-native/icon to v0.13.6
+
+## 0.5.10
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.1
+- Bump @fluentui-react-native/interactive-hooks to v0.18.2
+- Bump @fluentui-react-native/focus-zone to v0.10.8
+- Bump @fluentui-react-native/text to v0.14.1
+- Bump @fluentui-react-native/icon to v0.13.5
 
 ## 0.5.9