+++ packages/theming/default-theme/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/default-theme
 
-This log was last generated on Thu, 14 Jul 2022 18:09:51 GMT and should not be manually modified.
+This log was last generated on Mon, 29 Aug 2022 23:21:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.15.0
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Minor changes
+
+- Added unit tests (nkhalil942@gmail.com)
+- Bump @fluentui-react-native/theming-utils to v0.15.0
 
 ## 0.14.0